+++ server/api/admin/urunler/[id].delete.ts
@@ -1,8 +1,9 @@
 import { UrunlerService } from '../../../services/urunler.service';
-import { protect } from '../../../utils/protect';
+import { requirePermission } from '../../../utils/protect';
+import { PERMISSIONS } from '../../../utils/permissions';
 
 export default defineEventHandler(async (event) => {
-  await protect(event);
+  await requirePermission(event, PERMISSIONS.PRODUCTS.DELETE);
   const id = getRouterParam(event, 'id');
   if (!id) throw createError({ statusCode: 400, message: 'ID missing' });
   return await UrunlerService.delete(id);
